Correct serial drivers for invalid character count type.

+++ src/comm_drv_n0183_serial.cpp
@@ -416,7 +416,7 @@ void* CommDriverN0183SerialThread::Entry() {
     if (TestDestroy()) not_done = false;  // smooth exit
 
     uint8_t next_byte = 0;
-    size_t newdata = 0;
+    int newdata = 0;
     uint8_t rdata[2000];
 
     if (m_serial.isOpen()) {
@@ -552,7 +552,7 @@ void* CommDriverN0183SerialThread::Entry() {
     if (TestDestroy()) not_done = false;  // smooth exit
 
     uint8_t next_byte = 0;
-    size_t newdata = -1;
+    int newdata = -1;
     if (m_serial.isOpen()) {
       try {
         newdata = m_serial.read(&next_byte, 1);
